+++ apps/web/app/assets/LoginLogo.tsx
